RHTAP buildah task: add Syft SBOM generation
In order to add the SBOM generation, the previous build step needed to
be split up so that the image push is deferred to the end of the task.
It was also necessary to mount /var/lib/containers as a shared volume
between the steps.

After the changes, the task does the following:
- builds the image
- generate SBOMs for the built image and the source code
- merge SBOMs
- push SBOM
- push image

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>